+++ R/cran-app.R
@@ -16,6 +16,7 @@ make_dummy_package <- function(data, path) {
   unlink(package, recursive = TRUE)
   out <- dir()
   if (length(out) != 1) stop("Failed to build package ", package, " :(")
+  mkdirp(path)
   file.copy(out, path)
   out
 }
@@ -33,13 +34,35 @@ make_dummy_repo <- function(repo, packages = NULL, options = list()) {
     packages$Package <- paste0("pkg", seq_len(nrow(packages)))
   }
 
+  if (!"Version" %in% names(packages)) {
+    packages$Version <- "1.0.0"
+  } else {
+    packages$Version[is.na(packages$Version)] <- "1.0.0"
+  }
+
   dir_source <- utils::contrib.url("", "source")
   mkdirp(repo_source <- file.path(repo, dir_source))
 
-  packages$file <- character(nrow(packages))
+  extra <- packages
+  extra$file <- character(nrow(extra))
+
+  latest <- tapply(
+    extra$Version,
+    extra$Package,
+    function(x) as.character(max(package_version(x))),
+    simplify = TRUE
+  )
+  extra$archive <- latest[extra$Package] != extra$Version
+
   for (i in seq_len(nrow(packages))) {
-    fn <- make_dummy_package(packages[i, , drop = FALSE], repo_source)
-    packages$file[i] <- fn
+    pkg_dir <- if (extra$archive[i]) {
+      file.path(repo_source, "Archive", packages$Package[i])
+    } else {
+      repo_source
+    }
+
+    fn <- make_dummy_package(packages[i, , drop = FALSE], pkg_dir)
+    extra$file[i] <- fn
   }
 
   if (!isTRUE(options$no_packages)) {
@@ -56,18 +79,38 @@ make_dummy_repo <- function(repo, packages = NULL, options = list()) {
   }
 
   if (!isTRUE(options$no_metadata)) {
+    current <- extra[!extra$archive,, drop = FALSE]
     meta <- data.frame(
       stringsAsFactors = FALSE,
-      file = packages$file,
-      size = file.size(file.path(repo_source, packages$file)),
-      sha = unname(tools::md5sum(file.path(repo_source, packages$file))),
-      sysreqs = packages$SystemRequirements %||% rep(NA_character_, nrow(packages)),
-      built = if (nrow(packages)) NA_character_ else character(),
-      published = if (nrow(packages)) format(Sys.time()) else character()
+      file = current$file,
+      size = file.size(file.path(repo_source, current$file)),
+      sha = unname(tools::md5sum(file.path(repo_source, current$file))),
+      sysreqs = current$SystemRequirements %||% rep(NA_character_, nrow(current)),
+      built = if (nrow(current)) NA_character_ else character(),
+      published = if (nrow(current)) format(Sys.time()) else character()
     )
     outcon <- gzcon(file(file.path(repo_source, "METADATA2.gz"), "wb"))
     utils::write.csv(meta, outcon, row.names = FALSE)
     close(outcon)
+  }
+
+  if (!isTRUE(options$no_archive)) {
+    archive <- extra[extra$archive,, drop = FALSE]
+    adf <- list()
+    adir <- file.path(repo_source, "Archive")
+    if (file.exists(adir)) {
+      adirs <- dir(adir)
+      adf <- lapply(adirs, function(d) {
+        pkgs <- dir(file.path(adir, d), full.names = TRUE)
+        fi <- file.info(pkgs)
+        rownames(fi) <- basename(rownames(fi))
+        fi
+      })
+      names(adf) <- adirs
+    }
+
+    mkdirp(file.path(repo_source, "Meta"))
+    saveRDS(adf, file.path(repo_source, "Meta", "archive.rds"))
   }
 
   invisible()