+++ tests/testthat/test-2-metadata-api.R
@@ -3,6 +3,9 @@ pkgs <- dcf("
   Package: pkg1
   Version: 1.0.0
 
+  Package: pkg1
+  Version: 0.9.0
+
   Package: pkg2
   Version: 1.0.0
   Depends: pkg1